pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "paperlib"
version = "1.2.0"
requires-python = ">=3.14.3"
dependencies = [
    "click>=8.1",
    "pdfplumber>=0.11",
    "anthropic>=0.25",
    "python-dotenv>=1.0",
    "openai>=1.0",
]

[project.scripts]
paperlib = "paperlib.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff>=0.9",
    "mypy>=1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.14"
ignore_missing_imports = true
check_untyped_defs = true
